Update imports and .gitignore in Django FontAwesome5
The import for Django 'ugettext' was updated to 'gettext' in 'fields.py' to reflect the name change made in Django 3.0. This ensures the translation utility stays compatible with the newer versions of Django.

Additionally, a new folder '.idea' was added to the '.gitignore' file. This prevents the inclusion of JetBrains IDE settings in future commits, which is an overhead that is not necessary for the functioning of the project.