Remove setup.py, update pyproject.toml
Removed setup.py file as project is transitioning to use pyproject.toml for packaging and distribution configuration. Description of the package updated in pyproject.toml to match what was in the removed setup.py. Package metadata and resources inclusion also adjusted in the toml file for more current project state.