Add icon_as_html inclusion tag and adjust static file loading
Introduced an inclusion tag, "icon_as_html", to allow better customization of the FontAwesome icons in templating. Loading of the static files has been adjusted for better script organization. The ordering of the imported modules has been changed for better readability following PEP8. Reorganization of the raw code was made for better readability and easier maintenance of the code moving forward. Furthermore, the function "icon_as_html" has been outfitted with better typing and documentation for better understanding of the function's capability. The entirety of these changes improves the overall functionality and readability of the code, leading to better user experience when employing django-fontawesome-5.