Include html, js, css in package data of pyproject.toml
The pyproject.toml file in the django-fontawesome-5 project is modified to include ".html", ".js", and ".css" files under the package data. This ensures the included formats are correctly packaged during distribution. This makes our package more reliable as we ensure all necessary static files are included in the distribution and available for usage.