Updated icon prefix assignment in Fontawesome templates
In the fontawesome_5.py file, the method used to retrieve the predefined font icon prefix was revised. Before, it would use a static value from the settings module. Now it uses the get_prefix function from the app settings module, providing more flexibility when configuring the icons' prefix. This enhancement allows easier customization of icon prefixes and enhances the overall usability of the Django-fontawesome-5 package. This change improves the flexibility and ease of using and customizing Fontawesome icons in Django projects.